restructure and mentor comments
