+++ src/SearchResults.js
@@ -1,5 +1,7 @@
 import React from "react";
 import moment from "moment";
+import TableHead from "./TableHead";
+import TableRow from "./TableRow";
 
 function SearchResults(props) {
   return (
@@ -11,40 +13,11 @@ function SearchResults(props) {
             item.checkInDate,
             "days"
           );
-          return (
-            <tr key={index}>
-              <td scope="row">{item.id}</td>
-              <td>{item.title}</td>
-              <td>{item.firstName}</td>
-              <td>{item.surname}</td>
-              <td>{item.email}</td>
-              <td>{item.roomId}</td>
-              <td>{item.checkInDate}</td>
-              <td>{item.checkOutDate}</td>
-              <td>{nightsSpent}</td>
-            </tr>
-          );
+          return <TableRow item={item} key={index} nightsSpent={nightsSpent} />;
         })}
       </tbody>
     </table>
   );
 }
 
-function TableHead() {
-  return (
-    <thead className="thead-dark">
-      <tr>
-        <th scope="col">ID</th>
-        <th scope="col">Title</th>
-        <th scope="col">First Name</th>
-        <th scope="col">Surname</th>
-        <th scope="col">Email</th>
-        <th scope="col">Room ID</th>
-        <th scope="col">Check-In Date</th>
-        <th scope="col">Check-Out Date</th>
-        <th scope="col">No. Nights</th>
-      </tr>
-    </thead>
-  );
-}
 export default SearchResults;
