+++ src/TableHead.js
@@ -1,0 +1,21 @@
+import React from "react";
+
+function TableHead() {
+  return (
+    <thead className="thead-dark">
+      <tr>
+        <th scope="col">ID</th>
+        <th scope="col">Title</th>
+        <th scope="col">First Name</th>
+        <th scope="col">Surname</th>
+        <th scope="col">Email</th>
+        <th scope="col">Room ID</th>
+        <th scope="col">Check-In Date</th>
+        <th scope="col">Check-Out Date</th>
+        <th scope="col">No. Nights</th>
+      </tr>
+    </thead>
+  );
+}
+
+export default TableHead;
